+++ src/Common/ArgumentNullCheckAnalysis.cs
@@ -10,14 +10,15 @@ namespace Roslynator.CSharp;
 
 internal readonly struct ArgumentNullCheckAnalysis
 {
-    private ArgumentNullCheckAnalysis(ArgumentNullCheckStyle style, bool success)
+    private ArgumentNullCheckAnalysis(ArgumentNullCheckStyle style, string name, bool success)
     {
         Style = style;
+        Name = name;
         Success = success;
     }
 
     public ArgumentNullCheckStyle Style { get; }
-
+    public string Name { get; }
     public bool Success { get; }
 
     public static ArgumentNullCheckAnalysis Create(
@@ -37,6 +38,7 @@ internal readonly struct ArgumentNullCheckAnalysis
         if (statement is IfStatementSyntax ifStatement)
         {
             var style = ArgumentNullCheckStyle.None;
+            string identifier = null;
             var success = false;
 
             if (ifStatement.SingleNonBlockStatementOrDefault() is ThrowStatementSyntax throwStatement
@@ -52,22 +54,26 @@ internal readonly struct ArgumentNullCheckAnalysis
                 {
                     style = ArgumentNullCheckStyle.IfStatement;
 
-                    if (name is null
-                        || (nullCheck.Expression is IdentifierNameSyntax identifierName
-                            && string.Equals(name, identifierName.Identifier.ValueText, StringComparison.Ordinal)))
+                    if (nullCheck.Expression is IdentifierNameSyntax identifierName)
                     {
-                        if (semanticModel
-                            .GetSymbol(objectCreation, cancellationToken)?
-                            .ContainingType?
-                            .HasMetadataName(MetadataNames.System_ArgumentNullException) == true)
+                        identifier = identifierName.Identifier.ValueText;
+
+                        if (name is null
+                            || string.Equals(name, identifierName.Identifier.ValueText, StringComparison.Ordinal))
                         {
-                            success = true;
+                            if (semanticModel
+                                .GetSymbol(objectCreation, cancellationToken)?
+                                .ContainingType?
+                                .HasMetadataName(MetadataNames.System_ArgumentNullException) == true)
+                            {
+                                success = true;
+                            }
                         }
                     }
                 }
             }
 
-            return new ArgumentNullCheckAnalysis(style, success);
+            return new ArgumentNullCheckAnalysis(style, identifier, success);
         }
         else
         {
@@ -82,6 +88,7 @@ internal readonly struct ArgumentNullCheckAnalysis
         CancellationToken cancellationToken)
     {
         var style = ArgumentNullCheckStyle.None;
+        string identifier = null;
         var success = false;
 
         if (statement is ExpressionStatementSyntax expressionStatement)
@@ -97,16 +104,20 @@ internal readonly struct ArgumentNullCheckAnalysis
             {
                 style = ArgumentNullCheckStyle.ThrowIfNullMethod;
 
-                if (name is null
-                    || (invocationInfo.Arguments.SingleOrDefault(shouldThrow: false)?.Expression is IdentifierNameSyntax identifierName
-                        && string.Equals(name, identifierName.Identifier.ValueText, StringComparison.Ordinal)))
+                if (invocationInfo.Arguments.SingleOrDefault(shouldThrow: false)?.Expression is IdentifierNameSyntax identifierName)
                 {
-                    success = true;
+                    identifier = identifierName.Identifier.ValueText;
+
+                    if (name is null
+                        || string.Equals(name, identifierName.Identifier.ValueText, StringComparison.Ordinal))
+                    {
+                        success = true;
+                    }
                 }
             }
         }
 
-        return new ArgumentNullCheckAnalysis(style, success);
+        return new ArgumentNullCheckAnalysis(style, identifier, success);
     }
 
     public static bool IsArgumentNullExceptionThrowIfNullCheck(