+++ ChangeLog.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add analyzer "Simplify argument null check" ([RCS1255](https://github.com/JosefPihrt/Roslynator/blob/main/docs/analyzers/RCS1255.md)) ([#994](https://github.com/JosefPihrt/Roslynator/pull/994)).
   - Use `ArgumentNullException.ThrowIfNull` instead of `if` null check.
   - Not enabled by default
+- Add analyzer "Invalid argument null check" ([RCS1256](https://github.com/JosefPihrt/Roslynator/blob/main/docs/analyzers/RCS1256.md)) ([#888](https://github.com/JosefPihrt/Roslynator/pull/888)).
+  - This analyzer reports null checks of arguments that are:
+	- annotated as nullable reference type.
+	- optional and its default value is `null`.
 
 ### Changed